Replaced the artificial \baselinestretch linespacings with linespacing package.
It has the same effects, except it ignores certain environments. For example linespacing in footnotes is not doubled.
In the example document the only visible difference is that code listings no longer have double spacing.

Last but not least, the package is much more writer friendly, if someone wants to change spacing on the go.